me mande una cagadita y lo arregle

+++ main.c
@@ -6,14 +6,14 @@
 int main(char *argv[])
 {
     FILE* arch;
-    int16_t RAM[4096];
-    int16_t REG[16];
-    int16_t instruccion;
+    int RAM[4096];
+    int REG[16];
+    int instruccion;
     int i=0,DS;
 
     if ((arch = fopen(argv[1], "r")) == NULL)
         return 1;
-    while(fread(RAM[i],sizeof(int16_t),arch)!=NULL)
+    while(fread(RAM[i],sizeof(int),arch)!=NULL)
       i++;
     DS = i-1;
